Refactor(task): reorganise TaskList and improve code clarity
- Moved TaskList to the Task package for better logical grouping
- Added JavaDocs for enums in the Enums package to enhance documentation
- Updated comments for improved readability
- Removed unnecessary imports to clean up the codebase
