+++ src/main/java/command/ExitCommand.java
@@ -3,7 +3,7 @@ package command;
 import enums.CommandType;
 import enums.ExecutionStatus;
 import enums.OutputSymbol;
-import util.TaskList;
+import task.TaskList;
 
 /**
  * A command to terminate the application.
